+++ dandy/llm/config/config.py
@@ -1,12 +1,15 @@
 from abc import abstractmethod
 from typing_extensions import List, Union
 
-from pydantic import BaseModel, Field
+from pydantic import BaseModel, Field, field_validator
 
 from dandy.core.url import Url
+from dandy.llm.exceptions import LlmException
 from dandy.llm.service import Service
 from dandy.llm.request.request import BaseRequestBody
 
+_DEFAULT_CONTEXT_LENGTH = 4096
+_DEFAULT_MAX_COMPLETION_TOKENS = 512
 _DEFAULT_SEED = 77
 _DEFAULT_TEMPERATURE = 0.7
 _DEFAULT_CONNECTION_RETRY_COUNT = 10
@@ -20,6 +23,8 @@ class BaseLlmConfig(BaseModel):
     headers: Union[dict, None] = None,
     api_key: Union[str, None] = None,
     seed: Union[int, None] = _DEFAULT_SEED,
+    context_length: int = Field(_DEFAULT_CONTEXT_LENGTH, ge=0, le=8192)
+    max_completion_tokens: int = Field(_DEFAULT_MAX_COMPLETION_TOKENS, ge=0, le=2048)
     temperature: float = Field(_DEFAULT_TEMPERATURE, ge=0.0, le=1.0)
     connection_retry_count: int = Field(_DEFAULT_CONNECTION_RETRY_COUNT, ge=1, le=100)
     prompt_retry_count: int = Field(_DEFAULT_PROMPT_RETRY_COUNT, ge=1, le=10)
@@ -33,6 +38,8 @@ class BaseLlmConfig(BaseModel):
             query_parameters: Union[dict, None] = None,
             headers: Union[dict, None] = None,
             api_key: Union[str, None] = None,
+            context_length: int = _DEFAULT_CONTEXT_LENGTH,
+            max_completion_tokens: int = _DEFAULT_MAX_COMPLETION_TOKENS,
             seed: Union[int, None] = _DEFAULT_SEED,
             temperature: float = _DEFAULT_TEMPERATURE,
             connection_retry_count: int = _DEFAULT_CONNECTION_RETRY_COUNT,
@@ -48,6 +55,10 @@ class BaseLlmConfig(BaseModel):
         if api_key is not None:
             headers["Authorization"] = f"Bearer {api_key}"
 
+        self.validate_value(host, str)
+        self.validate_value(port, int)
+        self.validate_value(model, str)
+
         super().__init__(
             url=Url(
                 host=host,
@@ -59,6 +70,8 @@ class BaseLlmConfig(BaseModel):
             headers=headers,
             connection_retry_count=connection_retry_count,
             prompt_retry_count=prompt_retry_count,
+            context_length=context_length,
+            max_completion_tokens=max_completion_tokens,
             seed=seed,
             temperature=temperature,
             request_body=request_body,
@@ -77,6 +90,8 @@ class BaseLlmConfig(BaseModel):
     @abstractmethod
     def generate_request_body(
             self,
+            context_length: Union[int, None] = None,
+            max_completion_tokens: Union[int, None] = None,
             seed: Union[int, None] = None,
             temperature: Union[float, None] = None,
     ) -> BaseRequestBody:
@@ -84,12 +99,16 @@ class BaseLlmConfig(BaseModel):
 
     def generate_service(
             self,
+            context_length: Union[int, None] = None,
+            max_completion_tokens: Union[int, None] = None,
             seed: Union[int, None] = None,
             temperature: Union[float, None] = None,
     ) -> Service:
 
         return Service(
             self,
+            context_length=context_length,
+            max_completion_tokens=max_completion_tokens,
             seed=seed,
             temperature=temperature,
         )
@@ -97,3 +116,12 @@ class BaseLlmConfig(BaseModel):
     @property
     def service(self) -> Service:
         return self.generate_service()
+
+    def validate_value(self, value: Union[str, int], value_type: type):
+        exception_postfix = f'"{self.__class__.__name__}: {value}"'
+        if not isinstance(value, value_type):
+            raise LlmException(f'"{exception_postfix}" must be type {value_type.__name__}')
+        elif value is None:
+            raise LlmException(f'"{exception_postfix}: {value}" cannot be None')
+        elif value == '' or value == 0:
+            raise LlmException(f'"{exception_postfix}: {value}" cannot be empty')