+++ dandy/llm/config/openai.py
@@ -15,12 +15,15 @@ class OpenaiLlmConfig(BaseLlmConfig):
 
     def generate_request_body(
             self,
+            context_length: Union[int, None] = None,
+            max_completion_tokens: Union[int, None] = None,
+            seed: Union[int, None] = None,
             temperature: Union[float, None] = None,
-            seed: Union[int, None] = None
     ) -> BaseRequestBody:
 
         return OpenaiRequestBody(
             model=self.model,
+            max_completion_tokens=self.max_completion_tokens if max_completion_tokens is None else max_completion_tokens,
             seed=self.seed if seed is None else seed,
             temperature=self.temperature if temperature is None else temperature,
         )